- move decorateArea function to utils - load config for changeBgMarquee use-case
